(maint) Fix spec failure in hiera_spec try 2
Without this patch applied Puppet behavior tests are failing against
Hiera 1.0.0 because of the addition of a new, backwards-compatible,
default setting named `merge_behavior`.

This patch addresses the problem by only testing the configuration keys
we expect specific values for.  If the resulting configuration has
additional keys then we ignore them.

The related patch in Hiera is 4ac8372, "Add deep-merge feature to
backend lookups"

The related patch in Puppet that fails to pass against earlier versions
of Hiera is f980e17.