(maint) Fix spec failure in hiera_spec
Without this patch applied Puppet behavior tests are failing against
Hiera master because of the addition of a new, backwards-compatible,
default setting named `merge_behavior`.  This setting defaults to
`:native` and has no change in behavior in the default case.  The specs
are failing because they compare against the entire list of default
settings, comprehensively.

This patch addresses the problem by updating the list of default
settings to match Hiera master.

The related patch in Hiera is 4ac8372, "Add deep-merge feature to
backend lookups"