(yardoc) Add extras (private, DSL) to method details.
This adds the same button like extras for private and DSL for method
details as what is shown in the method list. 
This is good because it makes it easier to skip the non important
(private) methods and find the important ones (DSL) when browsing
through the method details.

The implementation overrides the method_signature template with one that
adds the extras. A setup script is also provided that defines a
helper method that is used in the template.