(maint) Update bundle dependencies
Without this patch applied the bundle dependencies for Facter and Hiera
are too conservative.  For example hiera 1.1.2 is currently released and
available via rubygems.org, but Bundler will never install this specific
version in our CI infrastructure because the specified dependency is '~>
1.0.0' which means greater than or equal to version 1.0.0 but less than
version 1.1.0.'  Similarly Facter 1.7 will never be used even though it
is backwards compatible and a valid satisfaction of the dependency
between Puppet 3 and Facter > 1.6.2.

This is a problem because Facter and Hiera follow semantic version
guidelines.  This means we should be able to depend on any minor version
of a particular major version and have some reasonable guarantee of
backwards compatibility.  settings to match Hiera master.

This patch simply updates the dependency specifications for all of the
libraries that are currently following semantic version guidelines.