(yardoc) Deprecate the use of :call option with :before and :after.
The implementation of properties/parameters with a :call option does not
honour the options :after and :before to make two calls (instead it
raises an exception). The documentation has been updated to reflect
this.

In this commit also several minor adjustments to the Property
documentation (typos, marking things as private/dsl).