(#17994) Rework symbolic mode acceptance test
Previously, the acceptance test was only testing that puppet could
modify the mode of an existing file/dir, but not when creating them. As
a result, #17994 went undetected.

This commit adds tests for that as well as adding test coverage for
other cases, such as the user 'a', which is short-hand for 'ugo', the
operation '-', permissions 's' and 't', and the absence of any
permission (o=).

In the process it uncovered a bug, when puppet applies ugo=rwxs to a
file, it results in 07777 instead of 06777. This is filed as #18355

One important thing to note about symbolic modes is that puppet always
uses a umask of 022 when applying the mode to file/directories that it
creates. When only a partial symbolic mode is specified, e.g. u=r, the
resulting mode will be 0444 (file) 0455 (directory), which may not be
what you expect. To ensure you don't accidentally grant more permissions
than intended (such as when setting root's umask to 077), it is
recommended that you fully specify the symbolic mode, e.g. 'u=r,go='

Paired-with: Andy Parker <andy@puppetlabs.com>