Maint: Remove example puppet-queue.conf file
This was not a Puppet configuration file; it was a stompserver config
file for use with puppetqd. Which is to say, it was a config file for a
third-party non-production-quality toy backend for a deprecated feature.
I am very sure that we don't need to keep it in the codebase.