Relax gem dependencies on Facter and Hiera
Without this patch applied an older version of Hiera is
automatically installed as a dependency installing Puppet.  This is a
problem because there are important bug fixes happening in later minor
versions of Hiera.  The same is true of Facter.  The cause of the
problem is a tight specification of `gem :hiera, "~> 1.0.0"` which means
"greater than 1.0.0 but less then 1.1.0".  1.1.2 is the current latest
bug fix version of Hiera.

This patch fixes the problem by relaxing the dependency to "~> 1.0"
which means "greater than 1.0.0 and less than 2.0.0"