Future proof authstore ruby issue 7477
Without this patch once a new patch level version of MRI comes out,
Travis CI will start failing because the test that fails on
`IPAddr.new('::2:3:4:5:6:7:8')` because of ruby issue 7477 [1].  This
patch fixes the problem by raising an exception if the bugged code is
still bugged.  When the issue is fixed the blocks of code that cause the
examples to be marked as pending will pass without exception.

[1]: http://goo.gl/Bb1LU