(Maint) Only show YARD docs for methods tagged with @api
Without this, docs are generated for any method with a comment before it. They
have the same appearance as methods marked public API. Rather than give the
appearance that these are public, this makes it so only methods that have been
given an explicit api classification are displayed. This will be very few at
first, but lets us grow our API specification in a more dependable way.

This also hides (void) on methods that don't return a meaningful value. Since
void is not a meaningful term in Ruby, this seemed to give the most intuitive
output.