(#11675) Refactor useradd provider
The useradd provider is a subclass of the nameservice provider and all
the property retrieval and property installation is handled in the
nameservice class.

This works for most properties but does not work for properties that can
be installed with usermod but cannot be retrieved with Etc.getpwnam. The
useradd provider defines special getter methods for these properties
(password_max_age, password_min_age, password, expiry). Because all of
these more or less look exactly the same we can remove duplicate code.