(#11675) Add value munging for nameservice provider
The nameservice provider defines a helper method `options` to specify
the flag for a property that will be used when creating the appropiate
useradd/usermod commandline. This commit now adds the options `munge` and
`umunge` to define methods that will automtically transformed the
property value when building the useradd/usermod command or when
retrieving the current value.

This munging is realized at a provider level so a unified value can be
used at a resource type level to represent different values at a
provider level.

One future use case could be the expiry property. We can define a value
of :absent at a resource type level to represent "account should not
expire" and represent it as '00-00-0000' in the pw provider and as ' '
in the useradd provider.