(#11675) Implement expiry in user provider
The useradd and pw provider claim to support expiry of user accounts.
And while they seem to be able to *set* the expiry, they don't seem to
implement any way to *get* the current value.

While this is still true for the pw provider (I am not familiar with
AIX) it should now work correctly with the useradd provider.

The expiry property uses munging at a provider level to convert
correctly to and from :absent (useraccount should not expire).