Maint: Deprecate dynamicfacts setting
The dynamicfacts setting was used to avoid recompiling catalogs when
only transient related facts were modified, e.g. free memory. According
to Luke, this optimization was removed when "we disabled staleness
checking for catalogs, which was early in the 0.24 release cycle", see
ticket #569.

This commit deprecates the setting rather than removing it, because it
the setting was actually used at one time, unlike `ca_md`.