(#18468) Clarify message in defaults for debian
The comment that tried to explain the START property did not clearly
align with the message generated by the installation:

  Setting up puppet (3.1.0-0.1rc1puppetlabs1) ...
  Starting puppet agent
  puppet not configured to start, please edit /etc/default/puppet to
  enable
  .

This changes the comment in the file to include the word "Enable" so
that a user can more confidently match it up to the "enable" word that
shows up in the message. It also explains the effect of the different
values.