(docs) Add COMMITTERS.md to guide committers
Without this patch the information a committer needs to possess in order
to merge contributions is spread all over the place.  The information is
also out of date.

This patch addresses the problem by consolidating the information into a
single place in the root of the project repository itself.
COMMITTERS.md should replace the following documents:

 * http://projects.puppetlabs.com/projects/puppet/wiki/Engineering_Process_Minimal_HOWTO
 * http://projects.puppetlabs.com/projects/puppet/wiki/Puppet_Design_Guidelines
 * http://projects.puppetlabs.com/projects/puppet/wiki/Internal_Development_Process
 * http://projects.puppetlabs.com/projects/puppet/wiki/Engineering_done_done

[ci skip]