Fix pson pure in 2.7.x for AIX.
Prior to this commit, PSON pure would fail on AIX for 2 reasons: 1)
AIX's Power processor is little endian and we specifically request
`utf16-be` for big endian, and 2) iconv on AIX is case-sensitive, which
means it knows about `UTF-8` and `UTF-16`, but not `utf-8` and `utf-16`.
This commit ignores the endian problem and instead just upcases the
requested encodings. This patch only applies to Puppet 2.7.x as iconv
has been removed from pson pure in 3.0+.