Use the right provider instance in launchd provider specs
These tests were creating a resource and a launchd provider instance
from that resource, when the resource already has a (default) provider instance.
In most cases this was working fine, because the resource's provider
instance wasn't often being referred to. However, a couple of tests were
trying to set the enable property of the resource, which fails on systems without an
enableable provider. Now, we explicitly set the resource's provider to
launchd, as use its provider instance for all of our tests.