(maint) RPM install should have --oldpackage
Prior to this commit, the RPM package provider would force the
`--oldpackage` option; however, older versions of RPM (i.e. 3.0.5 on
AIX) don't really like `--oldpackage` when you're doing an installation.
Rather than treat it as a NOOP, the older versions of RPM fail. This
commit changes the behaviour so that `--oldpackage` is only used when
the package is an up-/down-grade.