(#18458) Fix comment typos and removes unwanted required file.
This fixes typos in comments and instead of just copying the same text
in two location, there is now an updated explanation for what needs to
be checked later in the typeloader case. The two locations "compiler"
and "type_loader" now cross reference each other in the comments.

An unused require was also removed.