(AIX) Change ternary to unless for old RPMs
Prior to this commit, the RPM package provider used a less readable
ternary operator for determining which flags to pass to `rpm`. This
commit alters each of those to be a simple suffixed `unless` condition.

This commit also includes the more correct `rpm --version` output which
contains a "\n", and the corresponding `strip` on the version number.