(#18636) Create only one always acceptable Proc
The previous code would create a new Proc that returned true for every
instance of Token. This commit changes it so that only one is created.
Although there are a finite number of long lived tokens, extracting the
Proc out to a constant provides a name for it, which improves
readability, and reduces the number of objects in memory.