(AIX) Exclude arch from uninstall on old RPMs
Prior to this commit, the RPM provider assumed that `rpm` would
know what to do with the architecture (e.g. i386, x86_64, ppc)
when uninstalling a packages (e.g. `rpm -e mypkg.i386`).
Unfortunately, this only works for relatively modern versions of
`rpm` (4.2.1+).  This commit checks the version of `rpm` and will
add or exclude the architecture from the arguments sent to `rpm`.