(AIX) Tests for proper usage of --oldpackage w/ RPM
Prior to this commit, there were no tests for installing packages
with the RPM package provider. This commit adds tests to validate
that when a package doesn't already exist `rpm` is called with
`-i`, and when any version is already installed `rpm` is called
with `-U --oldpackage`. These tests are important because older
versions of RPM, such as 3.0.5 on AIX, barf when `--oldpackage` is
applied with `-i` whereas newer versions of RPM simply treat it as
a NOOP because it doesn't make sense with `-i`.