18030 - erb parse hiera results
The results provided by a hiera lookup are erb-parsed,
just like the inline_template or template functions.
This means that ruby can be used in your hiera yaml, but also
calls to puppet variables an even calls to puppet functions.
This results in the possibility to perform another hiera call.

The main code is located in hiera_puppet.rb.
The data_binding and indirector/hiera files have to be changed
due to the fact that the indirector didn't call HieraPuppet
but performed calls to hiera.lookup directly.

The change in the indirector/hiera file is written by ripienaar,
see: https://github.com/puppetlabs/puppet/pull/1217/files