(AIX) Convert RPM flags to array
Prior to this commit the flags passed to `rpm` during upgrade or
downgrade were a monolithic string of "-U --oldpackage". The
provider attempts to send this to RPM as a single flag, which
causes `rpm` to barf.  The commit changes the flags sent to `rpm`
to an array of strings, which allows upgrades and downgrades to
actually function as expected.

Sprintly: fix ticket:269, ticket:271