(AIX) Fix service resources on AIX (SRC provider)
Prior to this commit the SRC service provider was broken on recent
version of AIX (5.3, 6.1, 7.1). The SRC service provider was not able to
introspect running services nor was it able to enable/disable services.
This commit makes service introspection possible in addition to being
able to enable/disable specific services.