(#11276) Select tar command or minitar
Previously, the module tool relied on GNU versions of tar and gzip/gunzip
to build and install modules, which isn't portable on Windows in
particular.

This commit adds Gnu, Solaris, and Mini tar implementations for packing
and unpacking tar.gz format.

It modifies the module tool to select the appropriate implementation for
the platform it's currently running on.

Paired-with: Andy Parker <andy@puppetlabs.com>