Add SIGALRM handler for instant agent run
When puppet is configured with splay=true there is no easy
way to trigger an instant agent run.

To be backward compatible and not to overload peoples intrastructre
that depend on this delay, the agent uses SIGALRM for that.

To cause a instant run, type:

kill -SIGALRM `cat /var/run/puppet/agent.pid`